SPY-31749: xHE-AAC audio causes video frames to be dropped during seek
In a recent Exoplayer fix. An attempt to speed up seek in long mp3 segments, all the frames are dropped till the timestamp (with the assumption that every frame is SyncFrame for the following mimetypes).

AUDIO_AAC:
AUDIO_MPEG:
AUDIO_MPEG_L1:
AUDIO_MPEG_L2:

This breaks the use case for xHE-AAC (since in this new format not every frame is a SyncFrame). The problem manifests as

Dropped video frames in seek (i.e until the next xHE-AAC audio IPF is found).
Branched narrative titles transitions are no longer unstable.