improve thread warnings - ignore known usage of API's to limit future invalid usages
in general, we shouldn't be accessing exoplayer on any thread other than main

we have some use cases that are currently hard to get rid of (for example, ASE
accesses exoplayer to get its state during track selection, which *must* be on a
non-main thread)

previously, i just dropped these warnings to not pollute the log. instead, we will
add a list of allowed methods and ignore those. we can then trigger validation on
any unknown usages (for example, i found that we were preparing exoplayer
on a background thread when in branched playback. bugfix for that coming up)